modify startpoint double

+++ src/main/java/com/ssafy/ws/model/dto/Hotplace.java
@@ -19,6 +19,6 @@ public class Hotplace {
 	private String content;
 	private LocalDateTime createdAt;
 	private LocalDateTime updatedAt;
-	private Integer starPoint;
+	private double starPoint;
 	private byte[] image;
 }
